requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
ultralytics
PyMySQL
pydantic
email-validator
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
grpcio
grpcio-tools
httpx
rapidfuzz
